requirements.txt
flask==3.0.2
openai==1.12.0
pandas==2.2.1
python-dotenv==1.0.1
pyyaml==6.0.1
python-docx==1.1.0
scikit-learn==1.4.0
pydantic==2.6.3
openpyxl
flask-cors
